[cli] Lazily install and locally version localize
