+++ packages/labs/cli/package.json
@@ -81,6 +81,7 @@
     "@lit-labs/gen-utils": "^0.1.0",
     "@lit/localize-tools": "^0.6.1",
     "chalk": "^5.0.1",
+    "@lit-labs/gen-utils": "^0.0.1",
     "command-line-args": "^5.2.1",
     "command-line-commands": "^3.0.2",
     "command-line-usage": "^6.1.1",
